// proto/randomPackage/StreamMessage.ts
// Original file: proto/random.proto


export interface StreamMessage {
  'userId'?: (number);
  'userAvatar'?: (string);
  'message'?: (string);
  'userName'?: (string);
}

export interface StreamMessage__Output {
  'userId'?: (number);
  'userAvatar'?: (string);
  'message'?: (string);
  'userName'?: (string);
}